Hide imported modules (#38)
Move contents of __init__.py into _impl.py. Now __init__ only exports
the members of the public API.
