+++ tests/test_connection.py
@@ -10,7 +10,6 @@ from trio_websocket import (
     connect_websocket,
     connect_websocket_url,
     ConnectionClosed,
-    ListenPort,
     open_websocket,
     open_websocket_url,
     serve_websocket,
@@ -18,6 +17,7 @@ from trio_websocket import (
     wrap_client_stream,
     wrap_server_stream
 )
+from trio_websocket._impl import ListenPort
 
 
 HOST = '127.0.0.1'
